// src/app/model/files.ts
export class Salvataggio {
  descrizione: string;
  dataCreazione: Date;
  dataUltimoSalvataggio: Date;
  idCampionato: string;
  numeroTeams: number;
  stagione: string;
}

export class SalvataggioStagione {
  descrizione: string;
  dataCreazione: Date;
  dataUltimoSalvataggio: Date;
  idStagione: string;
  numeroCampionati: number;
  stagione: string;
}